chore(build-setup): rename lib.js to lib.ts
